Start work on adding a 'Future' continuation, which should run without the context being locked
Re-entrant streams are failing due to the need to lock the context while it's already locked: we need a type of continuation that runs out-of-context so things like this that might need to re-lock the context can work. We've already got some of the framework needed in the form of TalkContinuation::future and TalkContinuation::future_soon.