In the docs, change the way we add an inverted message to 'addInvertedMessage' from 'addInstanceMessage'
This change means that on Inverted types, 'addInstanceMessage' still does the thing it did before which is less confusing. It also fixes a problem where the 'addInstanceMessage' call is hard to override so we can register an inverted message type.

Think we still need a way to get the 'actual' class a message was sent to, which is potentially a bit weird with the current implementation of script classes (where we generate a TalkClass but represent it with an object instance)