Add tokio as a dependency
We'll make the initial version of the pipe code use tokio as we probably want to support it anyway.

One of the things about tokio is that it introduces a bunch of hidden state in the runtime which I kind of dislike, which in turn makes the possibility of implementing a 'raw' IO version tempting, but it's also a bunch of extra effort.