First attempt at implementing the 'stream_raw' function
This doesn't quite work, the problem is that we don't write to the output until the input stream is awaited which can lead to blocking. Also it's not entirely clear when the various streams are released. (If we could schedule a future elsewhere this might work, but that has lifetime issues)