Problem is that there's a filter on the 'All' input but not on the specific subprogram input.
I think it's unintuitive to have the 'all' filter not apply here as it makes general connections work but specific connections fail, so if we get a target that has a filter on it's 'all' input and we get a specific stream that it applies to, map to a filtered version instead of an unfiltered version.

This definitely produces another case that can be confusing where the user might expect there to be no filter for this
connection, but right now this seems like it will be pretty rare in general (particularly as there's ultimately only one
valid input type for a stream)

If the filter is used to redirect stuff away from a stream this currently won't happen either.