Bring over some of the core type
There's a bunch of things missing that I've commented out for the moment. One issue is what to do about the mutexes, we probably need a way to redefine these so we can switch between wasm with no threads and other things with multiple threads.

Could be an issue with BoxFutures too as they might need to turn into local futures...